split vanilla
