+++ vowpalwabbit/core/CMakeLists.txt
@@ -344,6 +344,7 @@ if(VW_BUILD_LARGE_ACTION_SPACE)
     ${vw_core_sources}
     src/reductions/cb/cb_explore_adf_large_action_space.cc
     src/reductions/cb/details/large_action/model_weight_rand_svd_impl.cc
+    src/reductions/cb/details/large_action/vanilla_rand_svd_impl.cc
     src/reductions/cb/cb_actions_mask.cc src/qr_decomposition.cc
   )
 endif()
